USer can now update there profile fully

+++ backend/app/Services/ProfileService.php
@@ -9,11 +9,16 @@ use App\Models\Artist;
 use App\Models\Collage;
 use App\Models\Piece;
 use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class ProfileService
 {
+
+    use AuthorizesRequests;
+
     /**
      * Create a new class instance.
      */
@@ -106,20 +111,32 @@ class ProfileService
 
     public function updateProfile($request)
     {
-        $profile = Auth::user()->profile;
         $validated = $request->validated();
 
+        if (!empty($validated['artist_id'])) {
+            $profile = Artist::find($validated['artist_id']);
+            Gate::authorize('updateArtist', $profile);
+        } else {
+            $profile = Auth::user()->profile;
+            Gate::authorize('update', $profile);
+        }
 
-        if ($request->hasFile('banner')) {
+        if ($request->hasFile('banner') && $profile->banner != null) {
             Storage::disk('public')->delete($profile->banner);
             $validated['banner'] = $request->file('banner')->store('banners', 'public');
         }
 
-        if ($request->hasFile('picture')) {
+        if ($request->hasFile('picture') && $profile->picture != null) {
             Storage::disk('public')->delete($profile->picture);
             $validated['picture'] = $request->file('picture')->store('pictures', 'public');
         }
 
+        unset($validated['artist_id']);
         $profile->update($validated);
+
+        return response()->json([
+            'success' => true,
+            'message'=> 'profile updated',
+        ]);
     }
 }
